move files into Analyser namespace
